MMfem: Finite Element Library

Welcome to the documentation page.

Note

Page under construction

Mesh generate

MMfem.rectangle_mesh — Method

rectangle_mesh(a,b,c,d,N)

This function generates a triangulation in the rectangle [a,b]x[c,d] using N subdivisions in each interval.

MMfem.nodes_tri — Method

nodes_tri(tri::Triangulation)

This function returns the array of nodes of the triangulation.

MMfem.elements_tri — Method

elements_tri(tri::Triangulation)

This function returns the numbering of each element of the triangulation.

MMfem.boundary_nodes — Method

boundary_nodes(tri::Triangulation)

This function returns the numbering of the edge nodes.

MMfem.mesh_data — Method

mesh_data(tri::Triangulation)

Plot functions

MMfem.plot_mesh — Method

plot_mesh(tri::DelaunayTriangulation)

Plot the mesh

MMfem.scalar_plot — Method

scalar_plot

This function graphs a scalar function.

MMfem.vectorial_plot — Method

vectorial_plot

This function graphs a vectorial function.

Base Functions

MMfem.β — Method

β(j,v,points,triangle)

Functions corresponding to the Lagrange basis

MMfem.∇β — Method

∇β(j,points,triangle)

Cuadratures

MMfem.quad — Method

quad(f,points,triangle;type="midpoints")

Operators

Stiffnes_Matrix

MMfem.Stiffnes_Δ — Method

Stiffnes_Δ(tri::Triangulation)

Load Vectors

MMfem.fβ — Method

fβ(f,j,v,points,triangle)

MMfem.LoadVector — Method

LoadVector(f,tri)

Boundary conditions

MMfem.dirichlet — Method

dirichlet(tri,dirichlet_nodes,uD)

Solver

MMfem.solver_problem — Method

solver_problem(A,F)