MMfem: Finite Element Library

Welcome to the documentation page.

Note

poner algo

Mesh generate

MMfem.rectangle_mesh — Method

rectangle_mesh

This function generates a mesh in the rectangle [a,b]x[c,d] using N subdivisions in each interval.

MMfem.nodes_tri — Method

nodes_tri

This function returns the array of nodes of the triangulation.

MMfem.elements_tri — Method

elements_tri

This function returns the numbering of each element of the triangulation.

MMfem.boundary_nodes — Method

boundary_nodes

This function returns the numbering of the edge nodes.

Plot functions

MMfem.scalar_plot — Method

scalar_plot

This function graphs a scalar function.

MMfem.vectorial_plot — Method

vectorial_plot

This function graphs a vectorial function.